+++ assets/meshes/buildings/base.mtl
@@ -1,0 +1,10 @@
+# Blender MTL File: 'allinoneblend.blend'
+# Material Count: 1
+
+newmtl None
+Ns 0
+Ka 0.000000 0.000000 0.000000
+Kd 0.8 0.8 0.8
+Ks 0.8 0.8 0.8
+d 1
+illum 2
