+++ assets/meshes/buildings/base.obj
@@ -1,0 +1,91 @@
+# Blender v2.79 (sub 0) OBJ File: 'allinoneblend.blend'
+# www.blender.org
+mtllib base.mtl
+o Cube.013_Cube
+v -0.330000 0.400000 0.330000
+v 0.000000 0.771250 0.330000
+v -0.330000 0.400000 -0.330000
+v 0.000000 0.771250 -0.330000
+v 0.330000 0.400000 0.330000
+v 0.330000 0.400000 -0.330000
+v -0.250000 0.000000 0.250000
+v -0.250000 0.000000 -0.250000
+v 0.250000 0.000000 0.250000
+v 0.250000 0.000000 -0.250000
+v -0.250000 0.400000 -0.250000
+v 0.250000 0.400000 -0.250000
+v -0.250000 0.400000 0.250000
+v 0.250000 0.400000 0.250000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 0.000000 0.000000
+vt 1.000000 0.000000
+vt 1.000000 1.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 0.000000 0.000000
+vt 1.000000 1.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 1.000000 0.000000
+vt 0.853553 0.853553
+vt 0.500000 0.000000
+vt 0.146447 0.146447
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 1.000000
+vt 1.000000 0.500000
+vt 0.853553 0.146447
+vt 0.000000 0.500000
+vt 0.146447 0.853553
+vt 0.500000 1.000000
+vn -0.7474 0.6644 0.0000
+vn 0.0000 0.0000 -1.0000
+vn 0.7474 0.6644 0.0000
+vn 0.0000 0.0000 1.0000
+vn 0.0000 -1.0000 0.0000
+vn -1.0000 0.0000 0.0000
+vn 1.0000 0.0000 0.0000
+usemtl None
+s off
+f 2/1/1 3/2/1 1/3/1
+f 3/4/2 4/5/2 6/6/2
+f 4/5/3 5/7/3 6/8/3
+f 5/9/4 2/1/4 1/10/4
+f 6/11/5 11/12/5 3/4/5
+f 13/13/6 8/14/6 7/15/6
+f 11/16/2 10/17/2 8/18/2
+f 12/19/7 9/20/7 10/21/7
+f 14/22/4 7/23/4 9/24/4
+f 10/25/5 7/23/5 8/18/5
+f 14/26/5 5/27/5 1/28/5
+f 2/1/1 4/29/1 3/2/1
+f 4/5/3 2/30/3 5/7/3
+f 6/11/5 12/31/5 11/12/5
+f 13/13/6 11/32/6 8/14/6
+f 11/16/2 12/31/2 10/17/2
+f 12/19/7 14/33/7 9/20/7
+f 14/22/4 13/34/4 7/23/4
+f 10/25/5 9/35/5 7/23/5
+f 14/26/5 12/36/5 6/37/5
+f 1/28/5 3/38/5 11/39/5
+f 1/28/5 11/39/5 13/40/5
+f 14/26/5 6/37/5 5/27/5
+f 1/28/5 13/40/5 14/26/5
