+++ src/entities/buildings/mod.rs
@@ -1,0 +1,3 @@
+mod base;
+
+pub use self::base::Base;
