+++ src/entities/mod.rs
@@ -1,1 +1,4 @@
-pub mod grid;
+pub mod buildings;
+mod tile;
+
+pub use self::tile::Tile;
